修改部分bug

+++ src/main/java/org/cjj/saaassistant/pojo/Order.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -13,7 +14,7 @@ import java.util.List;
 public class Order {
     private Integer id;
     private Integer userId; // 下单用户id
-    private Integer totalPrice; // 订单总价
+    private BigDecimal totalPrice; // 订单总价
     private String status; // 支付状态
     private Date createTime; // 创建时间
     private Date updateTime; // 更新时间
