+++ src/main/java/org/cjj/saaassistant/pojo/OrderProduct.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,7 +15,7 @@ public class OrderProduct {
     private Integer orderId; // 关联订单
     private Integer productId; // 关联商品
     private String productName;
-    private Integer productPrice;
+    private BigDecimal productPrice;
     private Integer quantity; // 购买数量
     private Date createTime; // 创建时间
     private Date updateTime; // 更新时间
